# By Andrey Derkach (3) and Sergey Reznik (3)
# Via Andrey Derkach (6) and others
* commit '44eeac50b3e6021e03d17ce2ce01b3af1209692d [formerly 0a3da00f142d2568ddc54a36fe611c52b5f15fc6]':
  Beautiful code in SpriteObject
  Redundant assert removed from SpriteObject
  [DF-8435] SpriteObject now does not saves render batches
  [DF-8361] Render initialization in the right way
  [DF-8361] fix of x64 warnings
  [DF-8361] fix of x64 warnings


Former-commit-id: 8843840563cebdf22fa727ca6a8e8b7cc61e04d0

+++ Sources/Internal/Render/RHI/Common/MCPP/expand.cpp
@@ -2367,7 +2367,7 @@ int* pragma_op /* Flag of _Pragma (not used in prestd)     */
         case NUM: /* Number token         */
         case OPE: /* Operator or punct.   */
         case NAM: /* Identifier           */
-            len = strlen(mp);
+            len = static_cast<int>(strlen(mp));
             mp += len;
             break;
         case SEP: /* Special character    */
